🔧🍎 Disable LTO on macOS arm64 builds (#29)
Testing in
[`mqt-core`](https://github.com/munich-quantum-toolkit/core/pull/1356)
has revealed that using LTO leads to problems when compiling with
different compilers than the ones being used to compile the static
libraries.

This PR disables the LTO option for macOS arm64, which was the only
setting where it was enabled.

This also directly prepares a release, which should allow us to test the
actions capabilities of picking the latest released version of a
particular LLVM tag.